+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## v0.3.0 – 20th February 2026
+
+✨ Features:
+
+- Add `Waveform` for generating x-axis values for a given channel.
+- Add `TimeWaveform` for when you know that your channels use time domain.
+- Add `Channel.Waveform()` method for retrieving waveform from channel.
+- Add `Properties.GetTimestamp()` to get property as timestamp.
+- Add `Channel.Unit()` to easily get the unit for a given channel from the standard `unit_string` property used by LabVIEW. Returns empty string if unit string property not found or not a string.
+
+💥 Breaking changes:
+
+- No
+
+🐛 Bug fixes:
+
+- Also no
+
+🦋 New bugs introduced:
+
+- Hopefully not
+
 ## v0.2.0 – 19th February 2026
 
 ✨ Features: